+++ test/fixtures/clickInfo.js
@@ -1,0 +1,125 @@
+const firefoxInfo = {
+  currentTab: {
+    id: 2,
+    index: 1,
+    windowId: 5,
+    highlighted: true,
+    active: true,
+    pinned: false,
+    status: "complete",
+    discarded: false,
+    incognito: false,
+    width: 1268,
+    height: 854,
+    lastAccessed: 1532606200506,
+    audible: false,
+    mutedInfo: {
+      muted: false
+    },
+    isArticle: false,
+    isInReaderMode: false,
+    url: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+    title: "Di6uEBuVsAEYVBw.jpg:orig (JPEG Image, 25 × 24 pixels)",
+    favIconUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig"
+  },
+  linkText: "",
+  now: "2018-07-26T11:56:42.642Z",
+  pageUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+  sourceUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+  url: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+  suggestedFilename: null,
+  context: "MEDIA",
+  menuIndex: "1",
+  comment: "",
+  modifiers: [],
+  legacyDownloadInfo: {
+    menuItemId: "save-in-1--.",
+    editable: false,
+    parentMenuItemId: "save-in-_-_-root",
+    mediaType: "image",
+    linkText: "",
+    linkUrl: "",
+    srcUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+    pageUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+    frameUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+    frameId: 0,
+    modifiers: []
+  },
+  naiveFilename: "Di6uEBuVsAEYVBw.jpg:orig",
+  filename: "Di6uEBuVsAEYVBw.jpg:orig",
+  initialFilename: "Di6uEBuVsAEYVBw.jpg:orig",
+  filenamePatterns: [
+    [
+      {
+        name: "filename",
+        value: {},
+        type: "MATCHER"
+      },
+      {
+        name: "sourceurl",
+        value: {},
+        type: "MATCHER"
+      },
+      {
+        name: "capture",
+        value: "filename",
+        type: "CAPTURE"
+      },
+      {
+        name: "into",
+        value: ":$1:",
+        type: "DESTINATION"
+      }
+    ]
+  ]
+};
+
+const chromeInfo = {
+  currentTab: {
+    active: true,
+    audible: false,
+    autoDiscardable: true,
+    discarded: false,
+    height: 748,
+    highlighted: true,
+    id: 1069,
+    incognito: false,
+    index: 1,
+    mutedInfo: {
+      muted: false
+    },
+    openerTabId: 1060,
+    pinned: false,
+    selected: true,
+    status: "loading",
+    title: "Di6uEBuVsAEYVBw.jpg:orig (25×24)",
+    url: "chrome://newtab/",
+    width: 1236,
+    windowId: 1059
+  },
+  now: "2018-07-26T12:07:55.529Z",
+  pageUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+  sourceUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+  url: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+  suggestedFilename: null,
+  context: "MEDIA",
+  menuIndex: "1",
+  comment: "",
+  legacyDownloadInfo: {
+    editable: false,
+    frameId: 0,
+    mediaType: "image",
+    menuItemId: "save-in-1--.",
+    pageUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig",
+    parentMenuItemId: "save-in-_-_-root",
+    srcUrl: "https://pbs.twimg.com/media/Di6uEBuVsAEYVBw.jpg:orig"
+  },
+  naiveFilename: "Di6uEBuVsAEYVBw.jpg:orig",
+  filename: "Di6uEBuVsAEYVBw.jpg_orig",
+  initialFilename: "Di6uEBuVsAEYVBw.jpg:orig"
+};
+
+module.exports = {
+  chromeInfo,
+  firefoxInfo
+};